% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nadaraya_Watson.R
\name{Nadaraya_Watson}
\alias{Nadaraya_Watson}
\title{input your arguments as such:
Nadaraya_Watson(user_data_x,user_data_y,user_input_h,user_input_arg)}
\usage{
Nadaraya_Watson(user_data_x, user_data_y, user_input_h, user_input_arg)
}
\description{
input your arguments as such:
Nadaraya_Watson(user_data_x,user_data_y,user_input_h,user_input_arg)
}
