% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parzen_Rosenblatt.R
\name{Parzen_Rosenblatt}
\alias{Parzen_Rosenblatt}
\title{input your arguments as such:
Parzen_Rosenblatt(user_data,user_input_h,user_input)
You may input NA for h and for user_input_arg in which case
h will be produced by: 1.06min(s, inter_q / 1.349) * n^(-1/5)
and the function will be evaluated at the data points}
\usage{
Parzen_Rosenblatt(user_data, user_input_h, user_input_arg)
}
\description{
input your arguments as such:
Parzen_Rosenblatt(user_data,user_input_h,user_input)
You may input NA for h and for user_input_arg in which case
h will be produced by: 1.06min(s, inter_q / 1.349) * n^(-1/5)
and the function will be evaluated at the data points
}
