% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Priestley_Chao.R
\name{Priestley_Chao}
\alias{Priestley_Chao}
\title{input your arguments as such:
Priestley_Chao(user_data_x,user_data_y,user_input_h,user_input_arg)}
\usage{
Priestley_Chao(user_data_x, user_data_y, user_input_h, user_input_arg)
}
\description{
input your arguments as such:
Priestley_Chao(user_data_x,user_data_y,user_input_h,user_input_arg)
}
